Report generating

+++ {{cookiecutter.project_slug}}/src/products/models.py
@@ -5,8 +5,14 @@ from app.base.models import BaseModel
 
 class Product(BaseModel):
     title = models.CharField(max_length=255)
-    cost = models.FloatField()
-    price = models.FloatField()
+    cost = models.DecimalField(
+        decimal_places=2,
+        max_digits=10,
+    )
+    price = models.DecimalField(
+        decimal_places=2,
+        max_digits=10,
+    )
     count = models.PositiveIntegerField()
 
     def __str__(self):
